Completed db readout, added deleterequest

+++ Pogserver/Pogserver/GivePLZ/Payloads/Requests/RemoveDataRequest.cs
@@ -10,8 +10,19 @@ namespace Pogserver.GivePLZ.Payloads.Requests
         public string VariableName { get; set; }
         public override string HandleRequest(APIManager.APIContext ctx)
         {
+            if (!Database.IsConfigured)
+            {
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine("Cannot access Database");
+                Console.ForegroundColor = ConsoleColor.White;
+                return "";
+            }
             var request = JsonSerializer.Deserialize<RemoveDataRequest>(ctx.input);
-            Database.ExecuteCommand("DELETE FROM " + request.Table + " WHERE" + request.VariableName + "=" + request.ID);
+            Console.WriteLine(request.ID);
+            Console.WriteLine(request.Table);
+            Console.WriteLine(request.VariableName);
+            Console.WriteLine("DELETE FROM " + request.Table + " WHERE " + request.VariableName + " = " + request.ID);
+            Database.ExecuteCommand("DELETE FROM " + request.Table + " WHERE " + request.VariableName + " = " + request.ID);
             return "";
         }
     }
